backend/requirements.txt
fastapi
uvicorn
requests
urequests
pydantic
sqlalchemy
python-dotenv
joblib
scikit-learn